app/requirements.txt
Flask
mysql-connector
simplejson
flask-mysql
pytz
pyqrcode
pypng